Make GenerateProgramOfSize accessible to other packages
